# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BAL Market Inventory Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List inventory items with optional filters
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
        - name: marketId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: condition
          in: query
          schema:
            type: string
        - name: reusable
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createItem
      tags: [items]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /items/bulk-import:
    post:
      operationId: bulkImportItems
      tags: [items]
      summary: Bulk import items from CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item with photos and history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItemDetail"
    put:
      operationId: updateItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /items/{id}/duplicate:
    post:
      operationId: duplicateItem
      tags: [items]
      summary: Duplicate an item to another market/season
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuplicateItemBody"
      responses:
        "201":
          description: Duplicated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /items/{id}/photos:
    post:
      operationId: addItemPhoto
      tags: [items]
      summary: Add a photo to an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPhotoBody"
      responses:
        "201":
          description: Added photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemPhoto"
  /items/{id}/photos/{photoId}:
    delete:
      operationId: deleteItemPhoto
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /items/{id}/history:
    get:
      operationId: getItemHistory
      tags: [items]
      summary: Get history log for an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemHistory"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        seasonId:
          type: integer
        seasonLabel:
          type: string
        marketId:
          type: integer
        marketName:
          type: string
        plannedQuantity:
          type: integer
        actualQuantity:
          type: integer
        unitType:
          type: string
        status:
          type: string
        condition:
          type: string
        reusable:
          type: boolean
        isGeneric:
          type: boolean
        supplierId:
          type: integer
        supplierName:
          type: string
        storageLocationId:
          type: integer
        storageLocationName:
          type: string
        notes:
          type: string
        thumbnailUrl:
          type: string
        lastUpdatedAt:
          type: string
        updatedBy:
          type: string
        createdAt:
          type: string
      required: [id, name, departmentId, seasonId, marketId, plannedQuantity, actualQuantity, unitType, status, condition, reusable, isGeneric, lastUpdatedAt, createdAt]
    InventoryItemDetail:
      allOf:
        - $ref: "#/components/schemas/InventoryItem"
        - type: object
          properties:
            photos:
              type: array
              items:
                $ref: "#/components/schemas/ItemPhoto"
            history:
              type: array
              items:
                $ref: "#/components/schemas/ItemHistory"
          required: [photos, history]
    CreateItemBody:
      type: object
      properties:
        name:
          type: string
        departmentId:
          type: integer
        seasonId:
          type: integer
        marketId:
          type: integer
        plannedQuantity:
          type: integer
        actualQuantity:
          type: integer
        unitType:
          type: string
        status:
          type: string
        condition:
          type: string
        reusable:
          type: boolean
        isGeneric:
          type: boolean
        supplierId:
          type: integer
        storageLocationId:
          type: integer
        notes:
          type: string
        updatedBy:
          type: string
      required: [name, departmentId, seasonId, marketId, plannedQuantity, actualQuantity, unitType, status, condition, reusable, isGeneric]
    UpdateItemBody:
      type: object
      properties:
        name:
          type: string
        departmentId:
          type: integer
        seasonId:
          type: integer
        marketId:
          type: integer
        plannedQuantity:
          type: integer
        actualQuantity:
          type: integer
        unitType:
          type: string
        status:
          type: string
        condition:
          type: string
        reusable:
          type: boolean
        isGeneric:
          type: boolean
        supplierId:
          type: integer
        storageLocationId:
          type: integer
        notes:
          type: string
        updatedBy:
          type: string
    DuplicateItemBody:
      type: object
      properties:
        targetMarketId:
          type: integer
        targetSeasonId:
          type: integer
      required: [targetMarketId, targetSeasonId]
    BulkImportBody:
      type: object
      properties:
        csvData:
          type: string
        seasonId:
          type: integer
        marketId:
          type: integer
      required: [csvData, seasonId, marketId]
    BulkImportResult:
      type: object
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [imported, failed, errors]
    AddPhotoBody:
      type: object
      properties:
        url:
          type: string
        caption:
          type: string
        photoType:
          type: string
      required: [url]
    ItemPhoto:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        url:
          type: string
        caption:
          type: string
        photoType:
          type: string
        createdAt:
          type: string
      required: [id, itemId, url, createdAt]
    ItemHistory:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        field:
          type: string
        oldValue:
          type: string
        newValue:
          type: string
        changedBy:
          type: string
        changedAt:
          type: string
      required: [id, itemId, field, newValue, changedAt]
...