# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BAL Market Inventory Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateItemBody {
  name: string;
  departmentId: number;
  seasonId: number;
  marketId: number;
  plannedQuantity: number;
  actualQuantity: number;
  unitType: string;
  status: string;
  condition: string;
  reusable: boolean;
  isGeneric: boolean;
  supplierId?: number;
  storageLocationId?: number;
  notes?: string;
  updatedBy?: string;
}