# lib/api-zod/src/generated/types/duplicateItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BAL Market Inventory Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DuplicateItemBody {
  targetMarketId: number;
  targetSeasonId: number;
}